Refactor ExtrasGenerator redesign code into ExtrasGeneratorV2
