+++ lib/pdf_fill/extras_generator.rb
@@ -2,18 +2,8 @@
 
 module PdfFill
   class ExtrasGenerator
-    HEADER_FONT_SIZE = 14.5
-    SUBHEADER_FONT_SIZE = 10.5
-
-    attr_reader :extras_redesign
-
-    def initialize(extras_redesign: false, form_name: nil, submit_date: nil, start_page: 1, sections: nil)
+    def initialize
       @generate_blocks = []
-      @extras_redesign = extras_redesign
-      @form_name = form_name
-      @submit_date = format_date(submit_date)
-      @start_page = start_page
-      @sections = sections
     end
 
     def create_block(value, metadata)
@@ -25,13 +15,13 @@ module PdfFill
         i = metadata[:i]
         prefix += " Line #{i + 1}" if i.present?
 
-        pdf.text("#{prefix}:", { style: extras_redesign ? :normal : :bold })
-        pdf.text(value.to_s, { style: extras_redesign ? :bold : :normal })
+        pdf.text("#{prefix}:", { style: :bold })
+        pdf.text(value.to_s, { style: :normal })
       end
     end
 
     def add_text(value, metadata)
-      unless text? || extras_redesign
+      unless text?
         @generate_blocks << {
           metadata: {},
           block: lambda do |pdf|
@@ -48,17 +38,6 @@ module PdfFill
 
     def text?
       @generate_blocks.size.positive?
-    end
-
-    def populate_section_indices!
-      return if @sections.blank?
-
-      @generate_blocks.each do |generate_block|
-        metadata = generate_block[:metadata]
-        if metadata[:top_level_key].present?
-          metadata[:section_index] = @sections.index { |sec| sec[:top_level_keys].include?(metadata[:top_level_key]) }
-        end
-      end
     end
 
     def sort_generate_blocks
@@ -84,42 +63,13 @@ module PdfFill
       pdf.font('Roboto')
     end
 
-    def set_header(pdf)
-      pdf.repeat :all do
-        bound_width = pdf.bounds.width / 2
-        location = [pdf.bounds.left, pdf.bounds.top]
-        write_header_main(pdf, location, bound_width, HEADER_FONT_SIZE)
-        if @submit_date.present?
-          location[0] += bound_width
-          write_header_submit_date(pdf, location, bound_width, HEADER_FONT_SIZE)
-        end
-        pdf.pad_top(2) { pdf.stroke_horizontal_rule }
-      end
-    end
-
-    def add_page_numbers(pdf)
-      pdf.number_pages('Page <page>',
-                       start_count_at: @start_page,
-                       at: [pdf.bounds.right - 50, 0],
-                       align: :right,
-                       size: 9)
-    end
-
-    def generate_pdf(file_path, generate_blocks)
+    def generate_pdf(file_path)
+      generate_blocks = sort_generate_blocks
       Prawn::Document.generate(file_path) do |pdf|
         set_font(pdf)
-        set_header(pdf) if extras_redesign
 
         render_pdf_content(pdf, generate_blocks)
-        add_page_numbers(pdf) if extras_redesign
       end
-    end
-
-    def render_new_section(pdf, section_index)
-      return unless @extras_redesign && @sections.present?
-
-      pdf.move_down(20)
-      pdf.text(@sections[section_index][:label], { size: 14 })
     end
 
     def render_pdf_content(pdf, generate_blocks)
@@ -145,42 +95,8 @@ module PdfFill
       folder = 'tmp/pdfs'
       FileUtils.mkdir_p(folder)
       file_path = "#{folder}/extras_#{SecureRandom.uuid}.pdf"
-      populate_section_indices!
-      generate_blocks = sort_generate_blocks
-      generate_pdf(file_path, generate_blocks)
+      generate_pdf(file_path)
       file_path
-    end
-
-    def write_header_main(pdf, location, bound_width, bound_height)
-      pdf.bounding_box(location, width: bound_width, height: bound_height) do
-        pdf.text("<b>ATTACHMENT</b> to VA Form #{@form_name}",
-                 align: :left,
-                 valign: :bottom,
-                 size: bound_height,
-                 inline_format: true)
-      end
-    end
-
-    def write_header_submit_date(pdf, location, bound_width, bound_height)
-      pdf.bounding_box(location, width: bound_width, height: bound_height) do
-        pdf.text("Submitted on VA.gov on #{@submit_date}",
-                 align: :right,
-                 valign: :bottom,
-                 size: SUBHEADER_FONT_SIZE)
-      end
-    end
-
-    # Formats the submit_date for the PDF header
-    def format_date(date)
-      return nil if date.blank?
-
-      return "#{date['month']}-#{date['day']}-#{date['year']}" if date.is_a?(Hash)
-      return date.strftime('%m-%d-%Y') if date.is_a?(Date)
-
-      Date.parse(date).strftime('%m-%d-%Y')
-    rescue
-      Rails.logger.error("Error formatting submit date for PdfFill: #{date}")
-      nil
     end
   end
 end
