+++ lib/pdf_fill/hash_converter.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require 'pdf_fill/extras_generator'
 require 'pdf_fill/form_value'
 
 # frozen_string_literal: true
